requirements.txt
accelerate==0.34.2
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
einops==0.8.0
filelock==3.16.1
fire==0.6.0
fonttools==4.54.1
fsspec==2024.9.0
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.2
medmnist==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
psutil==6.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sympy==1.13.3
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
transformers-stream-generator==0.0.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3

opencv-python~=4.10.0.84
rasterio~=1.4.0